* [Contrib] fix osX compiling of create_probe
git-svn-id: https://sampleicc.svn.sourceforge.net/svnroot/sampleicc/trunk@206 088ae1b8-34b2-e24d-a5c0-8cc67dc321ae

+++ Contrib/Mac_OS_X/create_probe/create_probe.cpp
@@ -118,7 +118,7 @@ createRGBBitmapContext(size_t width, size_t height, bool deep)
 {
   CGContextRef context;
   
-  float black[4] = { 0.0, 0.0, 0.0, 1.0 };
+  CGFloat black[4] = { 0.0, 0.0, 0.0, 1.0 };
   size_t bitsPerComponent = deep ? 32 : 8;
   // somewhat problematic: will fail for "thousands of colors" but who uses
   // that setting, these days?
